+++ src/services/locator.ts
@@ -30,15 +30,15 @@ import {
 import { sortBy, isEmpty } from "lodash";
 import registry, { readRawConfigurations } from "@/services/registry";
 import { inputProperties } from "@/helpers";
-import {
-  MissingConfigurationError,
-  NotConfiguredError,
-} from "@/services/errors";
 import { fetch } from "@/hooks/fetch";
 import { validateRegistryId } from "@/types/helpers";
 import { PIXIEBRIX_SERVICE_ID } from "@/services/constants";
 import { forbidContext } from "@/utils/expectContext";
 import { ExtensionNotLinkedError } from "@/errors/genericErrors";
+import {
+  MissingConfigurationError,
+  NotConfiguredError,
+} from "@/errors/businessErrors";
 
 const REF_SECRETS = [
   "https://app.pixiebrix.com/schemas/key#",
